# openapi.yaml
openapi: 3.0.2
info:
  title: KristyLessonRecords
  description: |
    ## Автоматическая система записи и обработки онлайн-уроков

    ### Возможности:
    - 🎥 Автоматическая запись уроков через Recall.ai
    - 📝 Транскрипция через AssemblyAI
    - 🤖 AI-обработка через YandexGPT
    - 📚 Генерация учебных материалов
  version: 1.0.0
  contact:
    name: KristyLessonRecords Support
    email: support@kristylessons.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://ai.dr-study.ru
    description: Local development server
  - url: https://api.kristylessons.com
    description: Production server
...
paths:
...
  /lessons/{lesson_id}:
    get:
      tags:
        - Lessons
      summary: Получить информацию об уроке
      description: Возвращает текущий статус и информацию об уроке
      operationId: get_lesson
      parameters:
        - required: true
          schema:
            title: Lesson Id
            type: string
            description: ID урока
            example: lesson_1234567890.123
          name: lesson_id
          in: path
      responses:
        "200":
          description: Информация об уроке
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResponse"
        "404":
          description: Урок не найден
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    LessonResponse:
      title: LessonResponse
      required:
        - id
        - status
        - meeting_url
        - lesson_type
        - created_at
      type: object
      properties:
        id:
          title: Id
          type: string
          description: Уникальный ID урока
          example: lesson_1234567890.123
        status:
          title: Status
          type: string
          description: Статус обработки урока
          example: recording
          pattern: ^(recording|transcribing|processing|completed|failed)$
        meeting_url:
          title: Meeting Url
          type: string
          description: URL встречи
          example: https://zoom.us/j/1234567890
        lesson_type:
          title: Lesson Type
          type: string
          description: Тип урока
          example: chinese
        created_at:
          title: Created At
          type: string
          description: Время создания записи
          format: date-time
        transcript_available:
          title: Transcript Available
          type: boolean
          description: Доступна ли транскрипция
          default: false
        materials_available:
          title: Materials Available
          type: boolean
          description: Доступны ли учебные материалы
          default: false
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string 